requirements.txt
Flask==3.1.1
gunicorn==23.0.0
python-dotenv==1.0.1
Flask-APScheduler==1.13.1
requests==2.32.3
kafka-python>=2.0

numpy>=2.2
joblib>=1.4

onnxruntime>=1.18
